lavague-integrations/contexts/lavague-contexts-groq/pyproject.toml
[tool.poetry]
name = "lavague-contexts-groq"
version = "0.1.0"
description = "Groq integration for lavague"
authors = ["lavague-ai"]
readme = "README.md"
license = "Apache-2.0"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: Apache Software License",
]
keywords = ["LAM", "action", "automation", "LLM", "NLP", "RAG", "selenium", "playwright"]
homepage = "https://lavague.ai"
repository = "https://github.com/lavague-ai/LaVague/"
documentation = "https://docs.lavague.ai/en/latest/"
packages = [{include = "lavague/"}]

[tool.poetry.dependencies]
python = "^3.8.1"
lavague-core = "^0.1.0"
llama-index-llms-groq = "^0.1.3"
llama-index-embeddings-openai = "^0.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
